Dev david (#8)
* first master commit (#3)

* first development commit

* New set

* test commit

* Updated readme

* added bat file for windows

* moved to traditional login

* updated json connection class

Deleted Hydra Platform JsonConnection class to completely sever ties
with HP.

* updated readme

* deleted personal Wing IDE project files

* updated readme

* updated readme

* updated readme

* updated readme

* updated readme

* fixed readme

* tweaked login and index nav

tweaked login
changed index navbar - will be different than main app navbar

* updated readme

* fixed login

minor fix to login

* general fixes to bootstrap

The bootstrap framework is improving.

* cleanup of views.py

Functions were moved from views.py to connections.py and conversions.py.
