+++ OpenAgua/connection.py
@@ -58,4 +58,54 @@ class JSONObject(dict):
     def __init__(self, obj_dict):
         for k, v in obj_dict.items():
             self[k] = v
-            setattr(self, k, v)    
+            setattr(self, k, v)
+            
+def get_project_by_name(conn, project_name):
+    return conn.call('get_project_by_name', {'project_name':project_name})
+
+def get_network_by_name(conn, project_name, network_name):
+    project = get_project_by_name(conn, project_name)
+    return conn.call('get_network_by_name', {'project_id':project.id, 'network_name':network_name})
+
+def get_network(conn, network_id):
+    return conn.call('get_network', {'network_id':network_id})    
+
+def get_coords(network):
+    coords = {}
+    for n in network['nodes']:
+        coords[n.id] = [float(n.x), float(n.y)] 
+    return coords
+
+# get shapes of type ftype
+def get_shapes(shapes, ftype):
+    return [s for s in shapes if s['geometry']['type']==ftype]
+
+def add_network(conn, project_name):
+    network = conn.call('add_network', {'net':{'nodes':[], 'links':[]}})       
+
+# use this to add shapes from Leaflet to Hydra
+def add_features(conn, network_id, shapes):
+
+    # modify to account for possibly no network... create network instead of add node
+
+    # convert geoJson to Hydra features & write to Hydra
+    points = get_shapes(shapes, 'Point')
+    polylines = get_shapes(shapes, 'LineString')    
+
+    if points:
+        nodes = make_nodes(points)
+        if nodes:
+            nodes = conn.call('add_nodes', {'network_id': network_id, 'nodes': nodes})
+    if polylines:
+        network = get_network(conn, network_id)
+        coords = get_coords(network)                   
+        links = make_links(polylines, coords)
+        if links:                         
+            links = conn.call('add_links', {'network_id': network_id, 'links': links})
+
+def get_features(network):
+    coords = get_coords(network)
+    nodes = nodes_geojson(network.nodes, coords)
+    links = links_geojson(network.links, coords)
+    features = nodes + links
+    return features
