+++ OpenAgua/static/js/network-editor.js
@@ -39,47 +39,6 @@ map.addControl(fullScreenButton);
 // feature creation & editing toolbar
 // need to revise this for multiple feature types
 // see https://github.com/Leaflet/Leaflet.draw/issues/265 and http://jsfiddle.net/jacobtoye/xgdV4/2/
-
-// add custom drawing tools based on template
-
-L.Draw.MarkerA = L.Draw.Marker.extend({
-    initialize: function (map, options) {
-        this.type = 'A';
-
-        L.Draw.Feature.prototype.initialize.call(this, map, options);
-    },
-
-    addHooks: function () {
-        L.Draw.Marker.prototype.addHooks.call(this);
-
-        if (this._map) {
-            this._tooltip.updateContent({ text: '' });
-        }
-    }
-});
-
-//L.DrawToolbar.include({
-    //getModeHandlers: function (map) {
-        //return [
-            //{
-                //enabled: true,
-                //handler: new L.Draw.MarkerA(map, { icon: new L.Icon.Default() }),
-                //title: 'Place reservoir marker'
-            //},
-            //{
-                //enabled: true,
-                //handler: new L.Draw.Marker(map, { icon: new L.Icon.Default() }),
-                //title: 'Place demand marker'
-            //},
-            //{
-                //enabled: true,
-                //handler: new L.Draw.Marker(map, { icon: new L.Icon.Default() }),
-                //title: 'Place inflow marker'
-            //}
-        //];
-    //}
-//});
-
 var drawControl = new L.Control.Draw({
     draw: {
         position: 'topleft'
@@ -98,6 +57,17 @@ drawControl.setDrawingOptions({
     polyline: { guideLayers: guideLayers, snapDistance: 15 },
 });
 
+// load existing network
+$( document ).ready(function() {
+  $.getJSON($SCRIPT_ROOT + '/_load_network', function(data) {
+    var currentItems_geoJson = JSON.parse(data.result.features);
+    currentItems.clearLayers();
+    currentItems.addData(currentItems_geoJson);
+    $("#load_status").text(data.result.status_message);
+  });
+});
+
+// create features
 map.on('draw:created', function (e) {
     var type = e.layerType,
         layer = e.layer;
@@ -111,59 +81,54 @@ map.on('draw:created', function (e) {
     $("#save_status").text('EDITS NOT SAVED!');
 });
 
-map.on('draw:edited', function (e) {
-    var layers = e.layers;
-    var countOfEditedLayers = 0;
-    layers.eachLayer(function(layer) {
-        countOfEditedLayers++;
-    });
-    console.log("Edited " + countOfEditedLayers + " layers");
-});
+//map.on('draw:edited', function (e) {
+    //var layers = e.layers;
+    //var countOfEditedLayers = 0;
+    //layers.eachLayer(function(layer) {
+        //countOfEditedLayers++;
+    //});
+    //console.log("Edited " + countOfEditedLayers + " layers");
+//});
 
-// button - load existing network, if any
-$('button#load_network').bind('click', function() {
-  $.getJSON($SCRIPT_ROOT + '/_load_network', function(data) {
-    var currentItems_geoJson = JSON.parse(data.result.features);
-    currentItems.clearLayers();
-    currentItems.addData(currentItems_geoJson);
-    $("#load_status").text(data.result.status_message);
+//// button - save edits
+//// if successful, need to update map with data.result.network_data
+//$('button#save_edits').bind('click', function() {
+  //var new_features = getJson(newItems);
+  //if (new_features.length > 0) {
+    //$.getJSON($SCRIPT_ROOT + '/_save_network', {new_features: new_features}, function(data) {
+      //var currentItems_geoJson = JSON.parse(data.result.features);
+      //currentItems.clearLayers();
+      //newItems.clearLayers();
+      //currentItems.addData(currentItems_geoJson);
+      //status_message = data.result.status_message;
+    //}
+    //else {
+      //status_message = 'No edits detected. Nothing saved.'
+    //};
+    //$("#save_status").text(status_method);
+  //});
+//});
+
+//// button - clear edits
+//$('button#clear_edits').bind('click', function() {
+  //newItems.eachLayer(function(layer) {
+    //newItems.removeLayer(layer);
+    //guideLayers.pop(layer);
+    //$("#save_status").text('Edits cleared.');
+  //});
+//});
+
+//// get shapes to add
+//var getJson = function(items) {
+    //var shapes = [];
+    //var layerJson;
     
-  });
-  return false;
-});
+    //items.eachLayer(function(layer) {
+        //layerJson = layer.toGeoJSON();
+        //shapes.push(layerJson);
+    //});
 
-// button - save edits
-// if successful, need to update map with data.result.network_data
-$('button#save_edits').bind('click', function() {
-  $.getJSON($SCRIPT_ROOT + '/_save_network', {new_features: getJson(newItems)}, function(data) {
-    var currentItems_geoJson = JSON.parse(data.result.features);
-    currentItems.clearLayers();
-    newItems.clearLayers();
-    currentItems.addData(currentItems_geoJson); 
-    $("#save_status").text(data.result.status_message);
-  });
-});
+    //var jsonshapes = JSON.stringify({shapes: shapes});
 
-// button - clear edits
-$('button#clear_edits').bind('click', function() {
-  newItems.eachLayer(function(layer) {
-    newItems.removeLayer(layer);
-    guideLayers.pop(layer);
-    $("#save_status").text('Edits cleared.');
-  });
-});
-
-// get shapes to add
-var getJson = function(items) {
-    var shapes = [];
-    var layerJson;
-    
-    items.eachLayer(function(layer) {
-        layerJson = layer.toGeoJSON();
-        shapes.push(layerJson);
-    });
-
-    var jsonshapes = JSON.stringify({shapes: shapes});
-
-    return jsonshapes;
-};
+    //return jsonshapes;
+//};
