+++ OpenAgua/templates/head_wrapper.html
@@ -9,10 +9,10 @@
 <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.2.0/css/font-awesome.min.css" />
 
 <!--Bootstrap-->
-<!--<link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">-->
-<!--<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>-->
-<link rel="stylesheet" href="{{ url_for('static', filename='lib/bootstrap-4.0.0-alpha.2/css/bootstrap.min.css') }}" />
-<script src="{{ url_for('static', filename='lib/bootstrap-4.0.0-alpha.2/js/bootstrap.min.js') }}"></script>
+<link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
+<script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js"></script>
+<!--<link rel="stylesheet" href="{{ url_for('static', filename='lib/bootstrap-4.0.0-alpha.2/css/bootstrap.min.css') }}" />-->
+<!--<script src="{{ url_for('static', filename='lib/bootstrap-4.0.0-alpha.2/js/bootstrap.min.js') }}"></script>-->
 
 <!--Bootstrap customizations-->
 <link rel="stylesheet" href="{{ url_for('static', filename='css/theme.css') }}" />
