+++ OpenAgua/templates/master.html
@@ -11,8 +11,8 @@
   
   <div class="row">
     {% if project_name %}
-      <div class="col-md-offset-2">
-        <p style="font-size:10px padding:5px">Project: {{project_name}}; Network: {{network_name}}</p>
+      <div class="col-md-offset-2" style="padding:10px">
+        <p style="font-size:12px">Project: {{project_name}}; Network: {{network_name}}</p>
       </div>
     {% endif %}
   </div>
