+++ OpenAgua/templates/navbar.html
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-fixed-top navbar-dark bg-inverse">
+<nav class="navbar navbar-fixed-top navbar-inverse">
   <a class="navbar-brand" href="/">OpenAgua DSS</a>
   <ul class="nav navbar-nav">
     <li class="nav-item active">
@@ -18,7 +18,7 @@
     </li>    
   </ul>
   <ul class="nav navbar-nav pull-right">
-    <li class="nav-item"><a class="nav-link" href="#"><span class="glyphicon glyphicon-user"></span> {{ username }}</a></li>
+    <li class="nav-item"><a class="nav-link" href="#"><span class="glyphicon glyphicon-user"></span> John</a></li>
     <li class="nav-item"><a class="nav-link" href="/logout"><span class="glyphicon glyphicon-log-in"></span> Logout</a></li>
   </ul>
