+++ OpenAgua/templates/navbar_index.html
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-fixed-top navbar-dark bg-inverse">
+<nav class="navbar navbar-fixed-top navbar-inverse">
   <a class="navbar-brand" href="/">OpenAgua DSS</a>
   <ul class="nav navbar-nav">
     <li class="nav-item">
