+++ OpenAgua/templates/network_editor.html
@@ -11,8 +11,8 @@
 <div id="map-edit" style="width: 800px; height: 400px; border: thin solid black"></div>
 <button type="button" class="btn btn-success" id="save_edits">Save</button>
 <button type="button" class="btn btn-danger" id="clear_edits">Clear edits</button>
-<div id="load_status"></div>
-<div id="save_status"></div>
+<div id="load_status" class="label label-success"></div>
+<div id="save_status" class="label label-danger"></div>
 
 <script src="{{ url_for('static', filename='js/network-editor.js') }}"></script>
