+++ run.py
@@ -1,2 +1,2 @@
 from OpenAgua import app
-app.run() 
+app.run()
